+++ src/models/stores/persistent-ui.ts
@@ -1,12 +1,8 @@
-import { getSnapshot, SnapshotIn, types } from "mobx-state-tree";
-import { debounce } from "lodash";
+import { getSnapshot, types } from "mobx-state-tree";
 import { AppConfigModelType } from "./app-config-model";
-import { kDividerHalf, kDividerMax, kDividerMin, UIDialogTypeEnum } from "./ui-types";
+import { kDividerHalf, kDividerMax, kDividerMin } from "./ui-types";
 import { WorkspaceModel } from "./workspace";
 import { DocumentModelType } from "../document/document";
-import { Logger } from "../../lib/logger";
-import { LogEventName } from "../../lib/logger-types";
-import { ITileModel } from "../tiles/tile-model";
 import { ENavTab } from "../view/nav-tabs";
 import { buildSectionPath, getCurriculumMetadata } from "../../../functions/src/shared";
 import { LearningLogDocument, LearningLogPublication, PersonalDocument,
@@ -15,37 +11,6 @@ import { LearningLogDocument, LearningLogPublication, PersonalDocument,
 import { UserModelType } from "./user";
 
 export const kSparrowAnnotationMode = "sparrow";
-
-type BooleanDialogResolver = (value: boolean | PromiseLike<boolean>) => void;
-type StringDialogResolver = (value: string | PromiseLike<string>) => void;
-let dialogResolver: BooleanDialogResolver | StringDialogResolver | undefined;
-
-// Information needed to scroll to a tile (for example, when a comment about a tile is selected)
-const ScrollToModel = types
-  .model("ScrollTo", {
-    tileId: types.string,
-    docId: types.string // key for user doc, path for problem doc
-  });
-
-export const UIDialogModel = types
-  .model("UIDialog", {
-    type: UIDialogTypeEnum,
-    text: types.string,
-    title: types.maybe(types.string),
-    className: "",
-    defaultValue: types.maybe(types.string),
-    rows: types.maybe(types.number)
-  })
-  .volatile(self => ({
-    promptValue: self.defaultValue
-  }))
-  .actions(self => ({
-    setPromptValue(value: string) {
-      self.promptValue = value;
-    }
-  }));
-type UIDialogModelSnapshot = SnapshotIn<typeof UIDialogModel>;
-type UIDialogModelSnapshotWithoutType = Omit<UIDialogModelSnapshot, "type">;
 
 // This generic model should work for both the problem tab, and the MyWork/ClassWork tabs
 // The StudentWorkspaces tab might work this way too. It has an open group which could be
@@ -60,35 +25,22 @@ export const UITabModel = types
     openSecondaryDocuments: types.map(types.string)
   });
 
-export const UIModel = types
-  .model("UI", {
-    annotationMode: types.maybe(types.string),
-    dividerPosition: kDividerHalf,
-    error: types.maybeNull(types.string),
-    activeNavTab: ENavTab.kProblems,
-    selectedTileIds: types.array(types.string),
-    selectedCommentId: types.maybe(types.string),
-    scrollTo: types.maybe(ScrollToModel),
-    showAnnotations: true,
-    showDemo: false,
-    showDemoCreator: false,
-    showTeacherContent: true,
-    showChatPanel: false,
-    dialog: types.maybe(UIDialogModel),
-    tabs: types.map(UITabModel),
-    problemWorkspace: WorkspaceModel,
-    learningLogWorkspace: WorkspaceModel,
-    teacherPanelKey: types.maybe(types.string),
-    dragId: types.maybe(types.string) // The id of the object being dragged. Used with dnd-kit dragging.
+export const PersistentUIModel = types
+  .model("PersistentUI", {
+    dividerPosition: kDividerHalf, //
+    activeNavTab: ENavTab.kProblems,// <-- keep this here
+    showAnnotations: true, //
+    showTeacherContent: true,//
+    showChatPanel: false, //
+    tabs: types.map(UITabModel), //
+    problemWorkspace: WorkspaceModel, //
+    teacherPanelKey: types.maybe(types.string), //
   })
   .volatile(self => ({
     defaultLeftNavExpanded: false,
     problemPath: ""
   }))
   .views((self) => ({
-    isSelectedTile(tile: ITileModel) {
-      return self.selectedTileIds.indexOf(tile.id) !== -1;
-    },
     get navTabContentShown () {
       return self.dividerPosition > kDividerMin;
     },
@@ -121,65 +73,6 @@ export const UIModel = types
     },
   }))
   .actions((self) => {
-    const alert = (textOrOpts: string | UIDialogModelSnapshotWithoutType, title?: string) => {
-      self.dialog = UIDialogModel.create(typeof textOrOpts === "string"
-                                          ? { type: "alert", text: textOrOpts, title }
-                                          : { type: "alert", ...textOrOpts });
-      dialogResolver = undefined;
-    };
-
-    const confirm = (textOrOpts: string | UIDialogModelSnapshotWithoutType, title?: string) => {
-      self.dialog = UIDialogModel.create(typeof textOrOpts === "string"
-                                          ? { type: "confirm", text: textOrOpts, title }
-                                          : { type: "confirm", ...textOrOpts });
-      return new Promise<boolean>((resolve, reject) => {
-        dialogResolver = resolve;
-      });
-    };
-
-    const prompt = (textOrOpts: string | UIDialogModelSnapshotWithoutType,
-                    defaultValue = "", title?: string, rows?: number) => {
-      self.dialog = UIDialogModel.create(typeof textOrOpts === "string"
-                                          ? { type: "prompt", text: textOrOpts, defaultValue, title, rows }
-                                          : { type: "prompt", ...textOrOpts });
-      return new Promise<string>((resolve, reject) => {
-        dialogResolver = resolve;
-      });
-    };
-
-    const resolveDialog = (value: string | boolean) => {
-      if (dialogResolver) {
-        dialogResolver(value as any);
-      }
-      closeDialog();
-    };
-
-    const closeDialog = () => {
-      self.dialog = undefined;
-      dialogResolver = undefined;
-    };
-
-    const setOrAppendTileIdToSelection = (tileId?: string, options?: {append: boolean}) => {
-      if (tileId) {
-        const tileIdIndex = self.selectedTileIds.indexOf(tileId);
-        const isCurrentlySelected = tileIdIndex >= 0;
-        const isExtendingSelection = options?.append;
-        if (isExtendingSelection) {
-          if (isCurrentlySelected) {
-            // clicking on a selected tile with a modifier key deselects it
-            self.selectedTileIds.splice(tileIdIndex, 1);
-          }
-          else {
-            self.selectedTileIds.push(tileId);
-          }
-        } else if (!isCurrentlySelected) {
-          self.selectedTileIds.replace([tileId]);
-        }
-        // clicking on an already-selected tile doesn't change selection
-      } else {
-        self.selectedTileIds.clear();
-      }
-    };
 
     const getTabState = (tab: string) => {
       let tabState = self.tabs.get(tab);
@@ -189,16 +82,10 @@ export const UIModel = types
       }
       return tabState;
     };
-
     return {
       alert,
       prompt,
-      confirm,
-      resolveDialog,
 
-      setAnnotationMode(mode?: string) {
-        self.annotationMode = mode;
-      },
       setDividerPosition(position: number) {
         self.dividerPosition = position;
       },
@@ -211,33 +98,11 @@ export const UIModel = types
       toggleShowChatPanel(show:boolean) {
         self.showChatPanel = show;
       },
-      setError(error: string) {
-        self.error = error ? error.toString() : error;
-        Logger.log(LogEventName.INTERNAL_ERROR_ENCOUNTERED, { message: self.error });
-        console.error(self.error);
-      },
-      clearError() {
-        self.error = null;
-      },
+
       setActiveNavTab(tab: string) {
         self.activeNavTab = tab;
       },
-      setSelectedTile(tile?: ITileModel, options?: {append: boolean}) {
-        setOrAppendTileIdToSelection(tile && tile.id, options);
-      },
-      setSelectedTileId(tileId: string, options?: {append: boolean}) {
-        setOrAppendTileIdToSelection(tileId, options);
-      },
-      removeTileIdFromSelection(tileId: string) {
-        self.selectedTileIds.remove(tileId);
-      },
-      setScrollTo(tileId: string, docId: string) {
-        self.scrollTo = ScrollToModel.create({ tileId, docId });
-      },
-      setShowDemoCreator(showDemoCreator: boolean) {
-        self.showDemoCreator = showDemoCreator;
-      },
-      closeDialog,
+
       rightNavDocumentSelected(appConfig: AppConfigModelType, document: DocumentModelType) {
         if (!document.isPublished || appConfig.showPublishedDocsInPrimaryWorkspace) {
           self.problemWorkspace.setAvailableDocument(document);
@@ -248,16 +113,18 @@ export const UIModel = types
             self.problemWorkspace.toggleComparisonVisible({override: true});
           }
           else {
-            alert("Please select a primary document first.", "Select Primary Document");
+            console.log("📁 persistent-ui.ts ------------------------");
+            console.log("\t🏭 rightNavDocumentSelected");
+            console.log("\t🔪 document:", document);
+            console.log("\t🥩 appConfig:", appConfig);
+            alert("Please select a primary document first.", /*"Select Primary Document"*/); //
           }
         }
       },
       setTeacherPanelKey(key: string) {
         self.teacherPanelKey = key;
       },
-      setDraggingId(dragId?: string) {
-        self.dragId = dragId;
-      },
+
       // We could switch this to openSubTab, however that would imply that the activeTab would be
       // switched if this is called. When all of the tabs are initialized this is called to setup
       // the default sub tab or each main tab, so we don't want to be switching the activeTab in
@@ -315,9 +182,6 @@ export const UIModel = types
     };
   })
   .actions(self => ({
-    clearSelectedTiles() {
-      self.selectedTileIds.forEach(tileId => self.removeTileIdFromSelection(tileId));
-    },
     /**
      * Update the top level tab in the resources panel (left side), and guess a sub tab to open to view
      * this document. Currently this only works with non curriculum docs.
@@ -356,7 +220,6 @@ export const UIModel = types
       }
       self.openSubTabDocument(navTab, subTab, doc.key);
     },
-
     openCurriculumDocument(docPath: string) {
       const {navTab, subTab} = getTabsOfCurriculumDoc(docPath);
       if (!subTab) {
@@ -368,16 +231,8 @@ export const UIModel = types
     }
 }));
 
-export type UIModelType = typeof UIModel.Type;
-export type UIDialogModelType = typeof UIDialogModel.Type;
+export type PersistentUIModelType = typeof PersistentUIModel.Type;
 
-export function selectTile(ui: UIModelType, model: ITileModel, isExtending?: boolean) {
-  ui.setSelectedTile(model, { append: !!isExtending });
-}
-
-// Sometimes we get multiple selection events for a single click.
-// We only want to respond once per such burst of selection events.
-export const debouncedSelectTile = debounce(selectTile, 50);
 
 // Maybe this should return the navTab and subTab
 export function getTabsOfCurriculumDoc(docPath: string) {
